requirements.txt
alabaster==0.7.12
aniso8601==7.0.0
asgiref==3.5.2
attrs==22.1.0
Babel==2.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
bumblebee-status==2.1.6
cairocffi==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==3.2.9
django-countries==7.4.2
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-stubs==1.13.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
entrypoints==0.4
fastjsonschema==2.16.2
graphene==2.1.9
graphene-django==2.15.0
graphene-stubs==0.15
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.3
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
mistune==2.0.4
msgpack==1.0.4
mypy==0.982
mypy-extensions==0.4.3
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nbsphinx==0.8.10
nest-asyncio==1.5.6
netifaces==0.11.0
packaging==22.0
pandocfilters==1.5.0
param==1.12.3
Pillow==9.2.0
platformdirs==2.6.0
promise==2.3
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pynvim==0.4.3
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python-xlib==0.31
pytz==2022.6
pyuca==1.2
pywayland==0.4.14
pywlroots==0.15.22
pyzmq==24.0.1
qtile==0.22.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-svg2pdfconverter==1.2.1
sqlparse==0.4.3
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12.1
typing_extensions==4.4.0
ueberzug==18.1.9
urllib3==1.26.13
webencodings==0.5.1
xcffib==0.12.1
xkbcommon==0.4
